refactor: reformat after PR
Update subproject `application` by linter rules.
