+++ application/src/main/kotlin/utils/placement_strategy/CircularPlacementStrategy.kt
@@ -1,6 +1,5 @@
 package utils.placement_strategy
 
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import themes.radiusVerticesStart
 import viewmodels.graphs.VertexViewModel
