+++ application/src/main/kotlin/views/MainScreen.kt
@@ -1,13 +1,14 @@
 package views
 
-import JetSettings
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.pager.VerticalPager
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
+import models.JetSettings
 import themes.MainTheme
 import utils.SideMenuTabType
 import views.pages.GraphViewPage
